+++ lib/cassandra/cluster.rb
@@ -41,8 +41,7 @@ module Cassandra
                    retry_policy,
                    address_resolution_policy,
                    connector,
-                   futures_factory,
-                   timestamp_generator)
+                   futures_factory)
       @logger                = logger
       @io_reactor            = io_reactor
       @executor              = executor
@@ -58,7 +57,6 @@ module Cassandra
       @address_resolver      = address_resolution_policy
       @connector             = connector
       @futures               = futures_factory
-      @timestamp_generator   = timestamp_generator
 
       @control_connection.on_close do |_cause|
         begin
@@ -206,8 +204,7 @@ module Cassandra
                            @retry_policy,
                            @address_resolver,
                            @connection_options,
-                           @futures,
-                           @timestamp_generator)
+                           @futures)
       session = Session.new(client, @execution_options, @futures)
       promise = @futures.promise