+++ spec/cassandra_spec.rb
@@ -373,20 +373,20 @@ describe Cassandra do
       expect(C.validate(synchronize_schema: 0)).to eq({synchronize_schema: true})
     end
 
-    it 'should massage :client_timestamps to a boolean' do
-      expect(C.validate(client_timestamps: nil)).to eq({client_timestamps: false})
-      expect(C.validate(client_timestamps: 1)).to eq({client_timestamps: true})
-      expect(C.validate(client_timestamps: 0)).to eq({client_timestamps: true})
-    end
+    it 'should massage :client_timestamps to a generator class or nil' do
+      expect(C.validate(client_timestamps: nil)).to eq({client_timestamps: nil})
 
-    it 'should validate :timestamp_generator' do
+      expected_class = RUBY_ENGINE == 'jruby' ?
+          Cassandra::TimestampGenerator::TickingOnDuplicate :
+          Cassandra::TimestampGenerator::Simple
+      expect(C.validate(client_timestamps: true)[:client_timestamps]).to be_instance_of(expected_class)
+      expect { C.validate(client_timestamps: Object.new) }.to raise_error(ArgumentError)
+
       valid_generator = Object.new
       def valid_generator.next
         42
       end
-      expect { C.validate(timestamp_generator: Object.new) }.to raise_error(ArgumentError)
-      expect(C.validate(timestamp_generator: valid_generator)).
-          to eq({ timestamp_generator: valid_generator })
+      expect(C.validate(client_timestamps: valid_generator)).to eq({ client_timestamps: valid_generator })
     end
 
     it 'should validate :connections_per_local_node' do